requirements.txt
streamlit
openai>=1.30
google-generativeai
python-dotenv